+++ src/Core/View.cs
@@ -12,8 +12,8 @@ namespace Blossom.Core
         public ElementTree Elements = new();
         public SKColor BackColor = SKColors.White;
 
-        public int Width => Browser.RenderRect.Width;
-        public int Height => Browser.RenderRect.Height;
+        public int Width => (int)Browser.RenderRect.Width;
+        public int Height => (int)Browser.RenderRect.Height;
 
         public event ForVoid Loop;
